Refactor venue-package name -> realization, move Competition
