+++ nva-datamodel-testutils/src/main/java/no/unit/nva/model/testing/PublicationInstanceBuilder.java
@@ -19,8 +19,8 @@ import no.unit.nva.model.contexttypes.place.UnconfirmedPlace;
 import no.unit.nva.model.instancetypes.artistic.Architecture;
 import no.unit.nva.model.instancetypes.artistic.ArchitectureOutput;
 import no.unit.nva.model.instancetypes.artistic.ArchitectureSubtypeEnum;
-import no.unit.nva.model.instancetypes.artistic.Competition;
-import no.unit.nva.model.instancetypes.artistic.venue.Venue;
+import no.unit.nva.model.instancetypes.artistic.realization.Competition;
+import no.unit.nva.model.instancetypes.artistic.realization.Venue;
 import no.unit.nva.model.instancetypes.PublicationInstance;
 import no.unit.nva.model.instancetypes.artistic.ArtisticDesign;
 import no.unit.nva.model.instancetypes.artistic.ArtisticDesignSubtype;
